requirements.txt
# Image IO & OCR
Pillow>=9.0.0
pytesseract>=0.3.10

# Numerical & filtering
numpy>=1.23.0
scipy>=1.7.0

# Google Drive API
google-api-python-client>=2.39.0
google-auth>=2.6.0
google-auth-httplib2>=0.1.0

# (Optional) if you use any extra auth helper libs:
# google-auth-oauthlib>=0.5.0
